Enable alternative epochs to 1970 to be used
In January 2038 the number of seconds since 1970-01-01 will overflow
the signed 32 bit integer commonly used for storing time values.  By
defining RTCX_EPOCH a year other than 1970 can be used for the epoch
and thus delay when the rollover occurs. If 1998 is used the overflow
is delayed until January 2066. The epoch selected must be 1970 or
differ from 1970 by a multiple of 28 years to ensure the assumptions
made for day of week are correct. The epoch must also be > 1900 and <
2100 for the same reasons.